Fix restorePlayingTweens (#146)
After loading a website using MoJS within an inactive tab, an error is thrown. 

"Cannot read property 'length' of undefined"

The error is due to the _savedTweens prop not being set unless savePlayingTweens has been called. This patch ensures that _savedTweens is always an array, preventing the error.

This bugs has been known of for a while, and has been fixed in dev branches, but for some reason this has not been patched on the master version, despite this being a massive issue for any production website, with an incredibly simple fix.